refac: optimize script for generating contract names
